+++ src/MetaheuristicsPlatform/Algorithms/Multiobjective/Knea/KneaReferences.cs
@@ -1,0 +1,12 @@
+using MetaheuristicsPlatform.Classification;
+namespace MetaheuristicsPlatform.Algorithms.Multiobjective.Knea;
+public static class KneaReferences
+{
+    public static ScientificReference ZhangTianJin2015 { get; } =
+        new(
+            "Zhang, Tian & Jin",
+            2015,
+            "A Knee Point Driven Evolutionary Algorithm for Many-Objective Optimization",
+            "IEEE Transactions on Evolutionary Computation 19(6), 761-776",
+            "10.1109/TEVC.2014.2378512");
+}